Made the _save_test_set_predictions save the items instead of the indexed items, since all the unknown items were becoming 0 and changing the ranking metrics. Also updated the evaluation to deal with it.

+++ src/mars_gym/evaluation/metrics/fairness.py
@@ -23,8 +23,8 @@ def calculate_fairness_metrics(
             sub_df = df[df[sub_key] == sub]
 
             y_true, y_pred = (
-                sub_df[ground_truth_key].astype(int),
-                sub_df[prediction_key].astype(int),
+                sub_df[ground_truth_key],
+                sub_df[prediction_key],
             )
             cnf_matrix = confusion_matrix(y_true, y_pred)
